Move hive.metastore-timeout to ThriftHiveMetastoreConfig
The config is applicable only when using `thrift` metastore.
