+++ presto-hive/src/test/java/io/prestosql/plugin/hive/TestHiveConfig.java
@@ -49,7 +49,6 @@ public class TestHiveConfig
                 .setAllowCorruptWritesForTesting(false)
                 .setPerTransactionMetastoreCacheMaximumSize(1000)
                 .setMetastoreSocksProxy(null)
-                .setMetastoreTimeout(new Duration(10, TimeUnit.SECONDS))
                 .setMinPartitionBatchSize(10)
                 .setMaxPartitionBatchSize(100)
                 .setMaxInitialSplits(200)
@@ -137,7 +136,6 @@ public class TestHiveConfig
                 .put("hive.allow-corrupt-writes-for-testing", "true")
                 .put("hive.per-transaction-metastore-cache-maximum-size", "500")
                 .put("hive.metastore.thrift.client.socks-proxy", "localhost:1080")
-                .put("hive.metastore-timeout", "20s")
                 .put("hive.metastore.partition-batch-size.min", "1")
                 .put("hive.metastore.partition-batch-size.max", "1000")
                 .put("hive.dfs.ipc-ping-interval", "34s")
@@ -222,7 +220,6 @@ public class TestHiveConfig
                 .setAllowCorruptWritesForTesting(true)
                 .setPerTransactionMetastoreCacheMaximumSize(500)
                 .setMetastoreSocksProxy(HostAndPort.fromParts("localhost", 1080))
-                .setMetastoreTimeout(new Duration(20, TimeUnit.SECONDS))
                 .setMinPartitionBatchSize(1)
                 .setMaxPartitionBatchSize(1000)
                 .setMaxInitialSplits(10)
