+++ presto-hive/src/test/java/io/prestosql/plugin/hive/metastore/thrift/TestThriftHiveMetastoreConfig.java
@@ -18,6 +18,7 @@ import io.airlift.units.Duration;
 import org.testng.annotations.Test;
 
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 import static io.airlift.configuration.testing.ConfigAssertions.assertFullMapping;
 import static io.airlift.configuration.testing.ConfigAssertions.assertRecordedDefaults;
@@ -30,6 +31,7 @@ public class TestThriftHiveMetastoreConfig
     public void testDefaults()
     {
         assertRecordedDefaults(recordDefaults(ThriftHiveMetastoreConfig.class)
+                .setMetastoreTimeout(new Duration(10, TimeUnit.SECONDS))
                 .setMaxRetries(9)
                 .setBackoffScaleFactor(2.0)
                 .setMinBackoffDelay(new Duration(1, SECONDS))
@@ -41,6 +43,7 @@ public class TestThriftHiveMetastoreConfig
     public void testExplicitPropertyMappings()
     {
         Map<String, String> properties = new ImmutableMap.Builder<String, String>()
+                .put("hive.metastore-timeout", "20s")
                 .put("hive.metastore.thrift.client.max-retries", "15")
                 .put("hive.metastore.thrift.client.backoff-scale-factor", "3.0")
                 .put("hive.metastore.thrift.client.min-backoff-delay", "2s")
@@ -49,6 +52,7 @@ public class TestThriftHiveMetastoreConfig
                 .build();
 
         ThriftHiveMetastoreConfig expected = new ThriftHiveMetastoreConfig()
+                .setMetastoreTimeout(new Duration(20, TimeUnit.SECONDS))
                 .setMaxRetries(15)
                 .setBackoffScaleFactor(3.0)
                 .setMinBackoffDelay(new Duration(2, SECONDS))
